PolyBot: Beta 0.1
In this release, a primitive version of PolyBot is presented. Right now, it can preform a basic spelling check, a Poly-Specific spelling check (for our building names and common terminology outside the dictionary), an advanced grammar check (provided by a external library which puts Grammarly to shame), and a LinkedIn parser for verification of student years, names, and positions. Note, for the position feature, an assumption of the immediately proceeding text providing context is used which will be refined in future revisions.