Additional Error Checking & Bug Fix
Non-existent LinkedIn profiles generating a search page Google Search result caused a bug in the web parsing. Thus, additional error checking ruling out search pages was preformed to avoid the error and to inform the end user that the LinkedIn profile does not exist for that specific name. While I was at it, I did some additional error checking in case Chromium's code interfacing decides to bug out on a weird popup or something during the web scraping to avoid crashing the bot.